refactor: Notification content 생성방식 변경
