+++ .github/workflows/amorgakco-cd.yml
@@ -17,5 +17,5 @@ jobs:
           script: |
             cd Backend
             git pull
-            sudo docker-compose down -v
+            sudo docker-compose down
             sudo docker-compose up -d --build
