+++ src/main/java/com/amorgakco/backend/notification/service/NotificationCreator.java
@@ -1,7 +1,8 @@
-package com.amorgakco.backend.notification.infrastructure.consumer;
+package com.amorgakco.backend.notification.service;
 
 import com.amorgakco.backend.member.domain.Member;
 import com.amorgakco.backend.notification.domain.NotificationTitle;
+import com.amorgakco.backend.notification.infrastructure.consumer.NotificationRequest;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -9,13 +10,17 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 public class NotificationCreator {
 
+    private static final String PARTICIPATION = "%s 님께서 모각코에 참여하길 원합니다.";
+    private static final String PARTICIPATION_APPROVE = "%s 님께서 %s 님의 모각코 참여를 허가했습니다.";
+    private static final String PARTICIPATION_REJECT = "%s 님께서 %s 님의 모각코 참여를 거절했습니다.";
+
     public static NotificationRequest participationNotification(
             final Member sender, final Member receiver) {
         return NotificationRequest.builder()
                 .sender(sender)
                 .receiver(receiver)
                 .notificationTitle(NotificationTitle.PARTICIPATION_REQUEST)
-                .content(sender.getNickname() + "님께서 모각코에 참여하길 원합니다.")
+                .content(PARTICIPATION.formatted(sender.getNickname()))
                 .build();
     }
 
@@ -26,7 +31,8 @@ public class NotificationCreator {
                 .receiver(receiver)
                 .notificationTitle(NotificationTitle.PARTICIPATION_APPROVED)
                 .content(
-                        sender.getNickname() + "님께서" + receiver.getNickname() + "의 모각코 참여를 허가했습니다.")
+                        PARTICIPATION_APPROVE.formatted(
+                                sender.getNickname(), receiver.getNickname()))
                 .build();
     }
 
@@ -37,7 +43,8 @@ public class NotificationCreator {
                 .receiver(receiver)
                 .notificationTitle(NotificationTitle.PARTICIPATION_REJECTED)
                 .content(
-                        sender.getNickname() + "님께서" + receiver.getNickname() + "의 모각코 참여를 거절했습니다.")
+                        PARTICIPATION_REJECT.formatted(
+                                sender.getNickname(), receiver.getNickname()))
                 .build();
     }
 }
